Use a fresh db session with every backend msg.
This fixes #93.  See that issue for a description of the problem.

This change will create a new session for every event handled by the
backend and will force a commit and close the session at the end of each
message.

For bonus points, we should employ the ZopeTransactionManager extension
in the future, but I'm not sure how it will play with the
flask_sqlalchemy extension.  Will check on that later.