Disable our usage of the koji proxyuser feature.
Unfortunately, the proxyuser method requires that we add all of the
modularity developers to the ProxyDNs entry in the staging koji-hub
config, and that's just not going to happen.

We get to re-work this in this ticket anyways,
https://pagure.io/fm-orchestrator/issue/112

While the proxyuser argument is in place, we can submit no builds to
koji at all (as far as I can tell), so this blocks further development.