New-style PDC deps.
This is only one half of the modulemd-1.0 changes needed around PDC.

This was done to get module handling their dependencies when they are
defined in terms of streams instead of NVRs.

In PDC we used to store the dep as one whole string, i.e.::

    base-runtime-0.25-11

Now we're going to store them with some more structure, since we can't
assume monotonically increasing version values::

    dependency: base-runtime
    stream: master

Pull-requests to PDC and PDC-updater are forthcoming.