Recover from errors with completed builds.
The base-runtime build from today had two of its 180 builds that
succeeded, but MBS failed to notice (due to an unhealthy way that I
killed it and restarted it earlier in the day).  The result was that the
component builds were done, but the module stayed stuck in the build
state even though its work was over.

This change alters the poller (which checks for builds that failed in
the buildSRPMFromSCM step which doesn't emit a message for our
consumer), to additionally check for builds that succeeded but which we
didn't notice.