Implement import content generator output into koji
This commit prepares a directory with outputs that will be uploaded to
koji (just the mmd file for now). It also adjusts koji_import to call
CGImport with the metadata and the directory to import the results.

Last step will be to call this method at the end of successful module
build