Restore the old message format.
The message format broke in 3cb41aa5dc2e11ff698cb2c3775ae96534d02dff

The resulted in tracebacks in pdc-updater when it tried to learn things
about a module that was just submitted.  It relied on certain fields
(`stream`) being in the message body.  This change restores those fields
to get things working again.