The Case of the Mysterious Dist Tag
We ran into a crazy bug in stg today where the `disttag` of every
component build .. for every module .. was the same!  (duh duh duh).

Turns out this had to do with the way we were writing the name, stream,
and version values back into the yaml file we recorded.  More correctly,
how we *weren't* writing those values back.