Revert "KojiContentGenerator: Convert MMD to UTF-8"
This reverts commit ae79b711d87a7f460fad256ba7ac81bd1c285b55.

This breaks MBS with python3 and fixes only single occurence of this
issue. We want to fix this in libmodulemd or find a better way how
to fix this to work on both python2 and python3.

We have libmodulemd PR open to address this issue on libmodulemd
level: https://github.com/fedora-modularity/libmodulemd/issues/184.